Move some ItemResource conversions away from VariantUtil.
VariantUtil may only contain modloader specific conversions.

ItemResource -> ItemStack with amount can be in ItemResource as it's platform neutral.

ItemStack -> ItemResource can be in VariantUtil as it's platform neutral.
