+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/util/VariantUtil.java
@@ -2,9 +2,7 @@ package com.refinedmods.refinedstorage2.platform.forge.util;
 
 import com.refinedmods.refinedstorage2.api.core.Action;
 import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
-import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
 
-import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import org.apache.logging.log4j.LogManager;
@@ -22,27 +20,12 @@ public final class VariantUtil {
 
     public static FluidStack toFluidStack(final FluidResource fluidResource, final long amount) {
         if (amount > Integer.MAX_VALUE) {
-            LOGGER.warn("Truncating too large amount to fit into FluidStack {}", amount);
+            LOGGER.warn("Truncating too large amount for {} to fit into FluidStack {}", fluidResource, amount);
         }
         return new FluidStack(fluidResource.fluid(), (int) amount, fluidResource.tag());
     }
 
     public static IFluidHandler.FluidAction toFluidAction(final Action action) {
         return action == Action.SIMULATE ? IFluidHandler.FluidAction.SIMULATE : IFluidHandler.FluidAction.EXECUTE;
-    }
-
-    // todo: does this need to move to common ?
-    public static ItemResource ofItemStack(final ItemStack itemStack) {
-        return new ItemResource(itemStack.getItem(), itemStack.getTag());
-    }
-
-    // todo: does this need to move to common ?
-    public static ItemStack toItemStack(final ItemResource itemResource, final long amount) {
-        if (amount > Integer.MAX_VALUE) {
-            LOGGER.warn("Truncating too large amount to fit into ItemStack {}", amount);
-        }
-        final ItemStack stack = new ItemStack(itemResource.item(), (int) amount);
-        stack.setTag(itemResource.tag());
-        return stack;
     }
 }
